requirements.txt
absl-py==0.7.1
argh==0.26.2
astor==0.7.1
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docker-pycreds==0.4.0
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
gpustat==0.5.0
gql==0.1.0
graphql-core==2.1
grpcio==1.20.1
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==3.0.5
numpy==1.16.3
nvidia-ml-py3==7.352.0
pathtools==0.1.2
Pillow==6.0.0
promise==2.2.1
protobuf==3.7.1
psutil==5.6.2
python-dateutil==2.8.0
PyYAML==5.1
redis==3.2.1
redlock==1.2.0
requests==2.21.0
Rx==1.6.1
scipy==1.2.1
sentry-sdk==0.7.14
shortuuid==0.5.0
six==1.12.0
smmap2==2.0.5
subprocess32==3.5.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.3
wandb==0.7.3
watchdog==0.9.0
Werkzeug==0.15.2

